% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_experience.R
\name{get_px_exp}
\alias{get_px_exp}
\title{Get patient experience data from MySQL (in the cloud)}
\usage{
get_px_exp(
  from = NULL,
  to = NULL,
  remove_demographics = TRUE,
  conn = conn_mysql_suce,
  open_data = TRUE,
  remove_optout = TRUE
)
}
\arguments{
\item{from}{String, specifying start date in "YYYY-MM-DD" format}

\item{to}{String, specifying end date in "YYYY-MM-DD" format}

\item{remove_demographics}{Logical, specifying whether to remove
demographic features (which must NOT be open sourced) (TRUE)
or leave them in (FALSE).}

\item{conn}{A database connection- by default done automatically}

\item{open_data}{Logical, indicating whether you want to use the pre-cleaned
open dataset that contains no demographic information or responses where the
respondent opted out of sharing. Setting to FALSE requires an authenticated
account}

\item{remove_optout}{Logical, specifying whether you want to remove people
who opted out of sharing. This MUST be done if data is shared. The open dataset
does not contain any opt outs, for this reason}
}
\value{

}
\description{
Note that you will need to configure Sys.getenv("UID_MYSQL_SUCE"),
and Sys.getenv("PWD_MYSQL_SUCE") or use open_data = TRUE to retrieve
the public dataset
}
\section{Last updated by}{

Chris Beeley
}

\section{Last updated date}{

2021-05-11
}

