% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_mysql.R
\name{connect_mysql}
\alias{connect_mysql}
\title{Connect to MySQL}
\usage{
connect_mysql(
  driver = "MySQL ODBC 8.0 Unicode Driver",
  server = "109.74.194.173",
  UID = "opendata",
  PWD = "letmein",
  Port = 3306,
  database = "SUCE",
  open_data = FALSE
)
}
\arguments{
\item{driver}{String, default is set to "MySQL ODBC 8.0 Unicode Driver"}

\item{server}{String, specifying the server address. The default value is correct-
an argument is provided in case we ever move the server}

\item{UID}{String, specifying the 'user id'. Defaults to "opendata" which is a special
user who can see the open data only. If you have your own account set it with
Sys.setenv("UID_MYSQL_SUCE" = "user.name")}

\item{PWD}{String, specifying SUPER secure personal password. Defaults to "letmein" which
is the open data user password You can set this using
Sys.setenv("PWD_MYSQL_SUCE" = "SECRET_PASSWORD")}

\item{Port}{Numeric, default is set to 3306}

\item{database}{String, default is set to "SUCE"}
}
\value{

}
\description{
MySQL needs to be installed, see https://dev.mysql.com/downloads/ for more information.
}
\section{Last updated by}{

Chris Beeley
}

\section{Last updated date}{

2021-04-03
}

